pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["noureddein <noureddein@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
django-phonenumber-field = "^6.3.0"
django-phonenumbers = "^1.0.1"
autopep8 = "^1.6.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
timedelta = "^2020.12.3"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
